requirements.txt
Django>=5.1
djangorestframework
